PROJETO: FINALIZANDO O BACKEND
 * Aplique o processo de Autorização em todos os endpoints de nossa API
 * Crie Presenters e Transformers em todos os repositories (deixe exibindo todos os dados por padrão - isso poderá ser mudado quando formos conversar com o Angular)

+++ app/Http/Controllers/ProjectMembersController.php
@@ -2,11 +2,19 @@
 
 namespace CodeProject\Http\Controllers;
 
+use CodeProject\Repositories\ProjectRepositoryEloquent;
 use CodeProject\Services\ProjectMembersService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
+use function dd;
 
 class ProjectMembersController extends Controller
 {
+
+    /**
+     * @var ProjectRepositoryEloquent
+     */
+    private $acl;
 
     /**
      * @var ProjectMembersService
@@ -17,8 +25,9 @@ class ProjectMembersController extends Controller
      *
      * @param ProjectMembersService $service
      */
-    public function __construct(ProjectMembersService $service) {
+    public function __construct(ProjectMembersService $service, ProjectRepositoryEloquent $acl) {
         $this->service = $service;
+        $this->acl = $acl;
     }
     
     /**
@@ -29,6 +38,9 @@ class ProjectMembersController extends Controller
      */
     public function create($id,$memberId)
     {
+        if($this->checkProjectPermissions($id) == false){
+            return ['error'=>'Access Forbidden'];
+        }
         $this->service->addMember($id,$memberId);
     }
 
@@ -39,6 +51,9 @@ class ProjectMembersController extends Controller
      * @return Response
      */
     public function show($id) {
+        if($this->checkProjectPermissions($id) == false){
+            return ['error'=>'Access Forbidden'];
+        }
         return $this->service->show($id);
     }
 
@@ -49,6 +64,9 @@ class ProjectMembersController extends Controller
      * @return Response
      */
     public function destroy($id,$memberId) {
+        if($this->checkProjectPermissions($id) == false){
+            return ['error'=>'Access Forbidden'];
+        }
         $this->service->removeMember($id,$memberId);
     }
     
@@ -61,4 +79,22 @@ class ProjectMembersController extends Controller
     public function is($id,$memberId) {
         dd($this->service->isMember($id,$memberId));
     }
+    
+    private function checkProjectOwner($projectId) {
+        $userId = Authorizer::getResourceOwnerId();  
+        return $this->acl->isOwner($projectId,$userId);
+    }
+    
+    private function checkProjectMember($projectId) {
+        $userId = Authorizer::getResourceOwnerId();       
+        return $this->acl->hasMember($projectId,$userId);
+    }
+    
+    private function checkProjectPermissions($projectId) {        
+        if($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)){
+            return true;
+        }
+        return false;
+    }
+    
 }
